+++ CHANGELOG.md
@@ -14,6 +14,7 @@ FEATURES:
 * Get/set basic settings
 * Get/set advanced settings
 * Set/delete staff data
+* Set fingerprint template
 
 BUG FIXES: